Made attributedMarkupString and attributedMarkdownString functions public
