+++ Marker/Classes/Marker.swift
@@ -1,9 +1,9 @@
 //
-//  MarkdownString.swift
+//  Marker.swift
 //  Marker
 //
-//  Created by Htin Linn on 5/4/16.
-//  Copyright © 2016 Prolific Interactive. All rights reserved.
+//  Created by Htin Linn on 5/15/17.
+//  Copyright © 2017 Prolific Interactive. All rights reserved.
 //
 
 import Foundation
@@ -14,8 +14,8 @@ import Foundation
 ///   - markdownText: String with Markdown tags.
 ///   - textStyle: Text style object containing style information.
 /// - Returns: Formatted Markdown text.
-internal func attributedMarkdownString(from markdownText: String,
-                                       using textStyle: TextStyle) -> NSAttributedString {
+public func attributedMarkdownString(from markdownText: String,
+                                     using textStyle: TextStyle) -> NSAttributedString {
     guard let (parsedString, elements) = try? MarkdownParser.parse(markdownText) else {
         return NSAttributedString(string: markdownText, textStyle: textStyle)
     }
@@ -49,6 +49,35 @@ internal func attributedMarkdownString(from markdownText: String,
                 attributedString.addAttributes([NSStrikethroughColorAttributeName: strikethroughColor],
                                                range: parsedString.range(from: element.range))
             }
+        }
+    }
+    
+    return attributedString
+}
+
+/// Returns formatted text as an attributed string with custom markup attributes applied.
+///
+/// - Parameters:
+///   - text: Text.
+///   - textStyle: Text style object containing style information.
+///   - markups: Markup information.
+/// - Returns: Formatted text.
+public func attributedMarkupString(from text: String,
+                                   using textStyle: TextStyle,
+                                   customMarkup markups: Markup) -> NSAttributedString {
+    guard
+        markups.count > 0,
+        let (parsedString, elements) = try? ElementParser.parse(text, for: markups.map { Symbol(character: $0.0) }) else {
+            return NSAttributedString(string: text, textStyle: textStyle)
+    }
+    
+    let attributedString = NSMutableAttributedString(string: textStyle.textTransform.applied(to: parsedString))
+    attributedString.addAttributes(textStyle.attributes,
+                                   range: NSRange(location: 0, length: parsedString.characters.count))
+    
+    elements.forEach { (element) in
+        if let markup = markups[Character(element.symbol.rawValue)] {
+            attributedString.addAttributes(markup.attributes, range: parsedString.range(from: element.range))
         }
     }
